Remove locale prefix from routes and navigation
