+++ frontend/__tests__/middleware.cookie.test.ts
@@ -1,12 +1,8 @@
 jest.mock('next/server', () => ({
   NextResponse: {
-    redirect: (url: URL) => {
+    next: () => {
       const cookies = new Map<string, { value: string }>();
       return {
-        headers: {
-          get: (name: string) =>
-            name.toLowerCase() === 'location' ? url.toString() : undefined,
-        },
         cookies: {
           get: (name: string) => cookies.get(name),
           set: (name: string, value: string) => {
@@ -20,18 +16,23 @@ jest.mock('next/server', () => ({
 
 const { middleware } = require('@/middleware');
 
-function createRequest() {
+function createRequest(hasCookie = false) {
   return {
-    url: 'https://example.com',
-    cookies: { get: () => undefined },
+    cookies: { get: () => (hasCookie ? { value: 'ru' } : undefined) },
   } as any;
 }
 
-describe('middleware locale redirect', () => {
-  it('always redirects to /ru and sets cookie', () => {
-    const request = createRequest();
+describe('middleware locale cookie', () => {
+  it('sets cookie when missing', () => {
+    const request = createRequest(false);
     const response = middleware(request);
-    expect(response.headers.get('location')).toBe('https://example.com/ru');
     expect(response.cookies.get('i18nextLng')?.value).toBe('ru');
   });
+
+  it('keeps existing cookie', () => {
+    const request = createRequest(true);
+    const response = middleware(request);
+    expect(response.cookies.get('i18nextLng')).toBeUndefined();
+  });
 });
+
