+++ frontend/app/__tests__/StatsPage.intim-poceluy.test.tsx
@@ -3,7 +3,7 @@ import i18n from "@/i18n";
 
 process.env.NEXT_PUBLIC_BACKEND_URL = "http://backend";
 
-const StatsPage = require("@/app/[locale]/stats/page").default;
+const StatsPage = require("@/app/stats/page").default;
 
 describe("StatsPage intim & poceluy", () => {
   beforeEach(async () => {
