+++ frontend/app/archive/[id]/page.tsx
