+++ frontend/app/games/[id]/page.tsx
