+++ frontend/app/stats/page.tsx
