+++ frontend/components/SettingsLink.tsx
@@ -2,17 +2,12 @@
 
 import Link from "next/link";
 import { useEffect, useState } from "react";
-import { usePathname } from "next/navigation";
 import { supabase } from "@/lib/supabase";
 import type { Session } from "@supabase/supabase-js";
 
 export default function SettingsLink() {
   const [session, setSession] = useState<Session | null>(null);
   const [isModerator, setIsModerator] = useState(false);
-  const pathname = usePathname();
-  const segments = pathname.split("/").filter(Boolean);
-  const lang = segments[0] ?? "ru";
-  const buildHref = (path: string) => `/${lang}${path === "/" ? "" : path}`;
 
   useEffect(() => {
     supabase.auth.getSession().then(({ data: { session } }) => {
@@ -40,5 +35,5 @@ export default function SettingsLink() {
 
   if (!isModerator) return null;
 
-  return <Link href={buildHref("/settings")}>Settings</Link>;
+  return <Link href="/settings">Settings</Link>;
 }
